CS/server/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
paho-mqtt==1.6.1
flask-cors==3.0.10
